Correciones bdd y otros
- Correciones bdd
- Nuevos Constructores vacíos
- Telefono de int a long
- Nombre bdd actualizado

+++ src/main/java/com/owo/TP_prg3/Clases/Proveedor/dto/FormProveedorDTO.java
@@ -12,6 +12,6 @@ public class FormProveedorDTO {
     private String nombreFantasia;
     private CondicionIVA condicion; 
     private Domicilio direccion; 
-    private Integer telefono; 
+    private Long telefono; 
     private String email;
 }
